+++ packages/ui/src/index.ts
@@ -3,6 +3,7 @@ export * from './box';
 export * from './button';
 export * from './form/primitives';
 export * from './icon';
+export * from './icon-button';
 export * from './stack';
 export * as Tooltip from './tooltip';
 export * from './visually-hidden';