+++ packages/dataviews/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Enhancements
+- DataViews: Add details form layout validation. [#74996](https://github.com/WordPress/gutenberg/pull/74996)
+
 ## 11.3.0 (2026-01-29)
 
 ### Enhancements